dynbrick: comments / typedefs.
Makes it easier to read, and harder to get confused about what
some things represent.  Should've done this a long while ago, was
always getting confused about what was indexing voxels vs. bricks
vs. whatever.

+++ DynamicBrickingDS.h
@@ -16,7 +16,7 @@ namespace tuvok {
 class DynamicBrickingDS : public FileBackedDataset {
 public:
   DynamicBrickingDS(std::shared_ptr<Dataset> ds,
-                    std::array<unsigned, 3> maxBrickSize);
+                    std::array<size_t, 3> maxBrickSize);
   virtual ~DynamicBrickingDS();
   virtual std::shared_ptr<const Histogram1D> Get1DHistogram() const;
   virtual std::shared_ptr<const Histogram2D> Get2DHistogram() const;
